feat(Monorepo): migrate to monorepo

+++ scripts/build/build-npm-package.js
@@ -1,9 +1,11 @@
 const fs = require('fs-extra');
 const path = require('path');
-const { writePackageJson } = require('./writePackageJson.plugin');
+const { createWritePackageHandler } = require('./writePackageHandler.plugin');
 
 const rootPath = process.cwd();
 const pkg = require(path.join(rootPath, 'package.json'));
+
+const licensePath = path.join(__dirname, '../../LICENSE');
 
 const targetPath = 'dist';
 const targetFolder = path.join(rootPath, targetPath);
@@ -11,40 +13,45 @@ const targetFolderES6 = path.join(targetFolder, 'es6');
 
 console.log('[Release]: prepare package start');
 
-const srcMovePaths = [
-  'README.md',
-  'LICENSE',
-  {
-    source: 'src/components/Dialer/DialPad/assets/RcDialerPadSounds.json',
-    to: 'RcDialerPadSounds.json',
-  },
-  { source: 'src/components/Icon/assets', to: 'icons' },
-  { source: 'src/components/Icon/assets', to: 'components/Icon/assets' },
-  { source: 'src/scss', to: 'scss' },
-];
+// write package.json
+const packageHandler = createWritePackageHandler(pkg, targetFolder);
+packageHandler.generateBundle();
 
-const subModules = ['icon'];
+const [moveConfigPath, ...args] = process.argv.slice(2);
 
-srcMovePaths.forEach((distMovePath) => {
-  if (typeof distMovePath === 'string') {
-    fs.copySync(
-      path.join(rootPath, distMovePath),
-      path.join(targetFolder, distMovePath),
-    );
-  } else {
-    fs.copySync(
-      path.join(rootPath, distMovePath.source),
-      path.join(targetFolder, distMovePath.to),
-    );
-    fs.copySync(
-      path.join(rootPath, distMovePath.source),
-      path.join(targetFolderES6, distMovePath.to),
-    );
+const configPath = moveConfigPath && path.join(rootPath, moveConfigPath);
+
+// copy license
+fs.copyFileSync(licensePath, path.join(targetFolder, 'LICENSE'));
+
+if (fs.existsSync(configPath)) {
+  console.log('[Release]: start copy files');
+  const npmPackageOptions = require(configPath);
+
+  // copy addition files
+  npmPackageOptions.movePaths.forEach((distMovePath) => {
+    if (typeof distMovePath === 'string') {
+      fs.copySync(
+        path.join(rootPath, distMovePath),
+        path.join(targetFolder, distMovePath),
+      );
+    } else {
+      fs.copySync(
+        path.join(rootPath, distMovePath.from),
+        path.join(targetFolder, distMovePath.to),
+      );
+      fs.copySync(
+        path.join(rootPath, distMovePath.from),
+        path.join(targetFolderES6, distMovePath.to),
+      );
+    }
+  });
+
+  console.log('[Release]: copy files completed');
+
+  if (npmPackageOptions.subModules) {
+    packageHandler.generateSubBundles(npmPackageOptions.subModules);
   }
-});
-
-const writePackageJsonHandler = writePackageJson(pkg, targetFolder);
-writePackageJsonHandler.generateBundle();
-writePackageJsonHandler.generateSubBundles(subModules);
+}
 
 console.log('[Release]: prepare package complete');
